Update settings to use new UI elements (#721)
This is a complete overhaul of how settings in the app are handled in
the new UI. Settings are migrated from `SharedPreferences` to a Proto
datastore which has many advantages.

As part of that, the settings page is rewritten to use the new UI
components. This is good too because it fixes some of the touch device
issues as well.

The downside is there is now a disconnect between the legacy UI settings
and new UI settings. Settings are migrated, but after that, most changes
made in one of the UI's settings will not be reflected in the other UI's
settings.