Support server custom localization setting (#722)
Adds support for the custom localization server setting. This allows for
replacing the strings used in the app.

See: https://docs.stashapp.cc/in-app-manual/interface/#custom-locales

The app uses server-side strings when possible, but there are
app-specific ones too. You can override the app-specific strings (such
as in
[`strings.xml`](https://github.com/damontecres/StashAppAndroidTV/blob/main/app/src/main/res/values/strings.xml))
by adding a JSON object with key `StashAppAndroidTV` to key-value pairs
of the android string names to your chosen values:
```json
{
  "actions": {
    "add_directory": "Some other description"
  },
  "StashAppAndroidTV": {
    "play_transcoding": "My custom play transcoding"
  }
}
```