Fix playback activity tracking during playlists (#731)
A user reported this on discord. If activity tracking is enabled, when
playing a playlist of scenes, as you moved to the next scene, the
activity tracker would report activity for the previous scenes when it
should have stopped.

The actual issue was that `trackActivityListener` was not stored in a
`mutableStateOf`, so re-compositions destroyed the reference. Then when
going to teh next scene, the code to cancel tracking for the previous
scene thought there was no `trackActivityListener` since the reference
was lost.

This PR moves the `trackActivityListener` into the view model which more
accurately follows the lifecycle of the entire playlist and movement
between scenes.